Add files via upload
I will give you a whole design of the web pages and some changes I request from you. Go to the attached file and based on the below order do those changes: Can you make some changes: First of all make the font just in one color in home page the "Creating Value for Homes & Properties" all in gold color and also Luxend in top in blue color of "X" and the logo should change as the attached logo with curved corners then: 1- add the attached logo to the pages for the Luxend brand (first attachment) 2- in quality and innovation, use the second attachment instead of the current one. 3- In our gallery page, change the photo of color and the premium painting to the third attachment. 4- in Geotechnical slope solutions, use the fourth attachment image. 5- for luxury landscape design in the gallery page use the fifth attachment image. 6- for land development and subdivision, part use the sixth attached image.